Implement ProPay into buy function (My Ky, Maren)

+++ src/main/java/de/hhu/sharing/web/SellableItemController.java
@@ -1,8 +1,12 @@
 package de.hhu.sharing.web;
 
+import de.hhu.sharing.data.TransactionPurchaseRepository;
 import de.hhu.sharing.model.SellableItem;
 import de.hhu.sharing.model.User;
+import de.hhu.sharing.propay.TransactionPurchase;
+import de.hhu.sharing.services.ProPayService;
 import de.hhu.sharing.services.SellableItemService;
+import de.hhu.sharing.services.TransactionPurchaseService;
 import de.hhu.sharing.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -23,6 +27,15 @@ public class SellableItemController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private ProPayService proService;
+
+    @Autowired
+    private TransactionPurchaseService transPurService;
+
+    @Autowired
+    private TransactionPurchaseRepository transPurRepo;
 
     @GetMapping("/sellItemDetails")
     private String sellItemDetails(Model model, @RequestParam("id") Long id){
@@ -91,6 +104,14 @@ public class SellableItemController {
         User owner = sellableItem.getOwner();
 
         // TODO something brilliant
+
+        if(!proService.enoughCredit(buyer,sellableItem)){
+            return "buy";
+        }
+
+        transPurService.createTransactionPurchase(sellableItem, owner, buyer);
+        TransactionPurchase transPur = transPurRepo.findByItemId(sellableItem.getId());
+        proService.initiateTransactionPurchase(transPur);
 
         sellableItemService.delete(id);
